// packages/uns-core/src/uns/process-config.d.ts
export declare const PROCESS_NAME_FILE: string;
export declare const PACKAGE_JSON_PATH: string;
interface PackageInfo {
    name: string;
    version: string;
}
export declare const PACKAGE_INFO: PackageInfo;
export declare const MQTT_UPDATE_INTERVAL = 10000;
export declare const ACTIVE_TIMEOUT = 10000;
export {};